Add registration and retrieval of single Blueprints with tests

+++ blueprint/src/main/kotlin/no/eolseng/pg6102/blueprint/db/BlueprintRepository.kt
@@ -2,4 +2,4 @@ package no.eolseng.pg6102.blueprint.db
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface BlueprintRepository : JpaRepository<Blueprint, Long>
+interface BlueprintRepository : JpaRepository<Blueprint, Int>
